added pagination on product service

+++ src/main/java/com/ecommerce/project/payload/CategoryResponse.java
@@ -15,7 +15,7 @@ public class CategoryResponse {
     private List<CategoryDTO> content;
     private int pageNumber;
     private int pageSize;
-    private int totalElements;
+    private long totalElements;
     private int totalPages;
     private boolean lastPage;
 }
